stock-market-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-market-mcp"
version = "0.1.0"
description = "Stock Market MCP Server - A comprehensive Model Context Protocol server for stock market data and portfolio management"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "yfinance>=0.2.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "asyncio",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

market-data = [
    "numpy>=1.24.0",
    "seaborn>=0.12.0",  # For better chart styling
]

database = [
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.28.0",  # for PostgreSQL
    "aiosqlite>=0.19.0",  # for SQLite
]

[project.urls]
Homepage = "https://github.com/yourusername/stock-market-mcp"
Repository = "https://github.com/yourusername/stock-market-mcp"
Issues = "https://github.com/yourusername/stock-market-mcp/issues"

[project.scripts]
stock-market-mcp = "stock_market_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["stock_market_mcp*"]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
